+++ internal/config.go
@@ -1,4 +1,4 @@
-package app
+package internal
 
 import (
 	"gopkg.in/yaml.v2"
