--- internal/domain/constants.go
@@ -1,8 +1,0 @@
-package domain
-
-type AccessState uint
-
-const (
-	NotRegistered AccessState = iota
-	Registered
-)
