--- internal/domain/source_connection.go
@@ -1,15 +1,0 @@
-package domain
-
-import (
-	"github.com/google/uuid"
-	"time"
-)
-
-type SourceConnection struct {
-	Gid        uuid.UUID  `json:"gid"`
-	AccountGid uuid.UUID  `json:"account_gid"`
-	SourceType SourceType `json:"source_type"`
-	UserId     string     `json:"user_id"`
-	CreateTs   time.Time  `json:"create_ts"`
-	UpdateTs   time.Time  `json:"update_ts"`
-}
